going for monorepo structure, reorganize repo
