+++ src/CLI.js
@@ -10,9 +10,9 @@ const execute = require('util').promisify(require('child_process').execFile);
 
 // local
 const { printDisplayFreq, end, execOptions } = require('./Globals');
-const CAR = require('./CmdArgsReader'); // 🚗
+const CAR = require('./car/CmdArgsReader'); // 🚗
 
-const Print = require('./Print')(printDisplayFreq);
+const Print = require('./pretty/Print')(printDisplayFreq);
 Print.setPrettyLabel('makever', 'yellow.black', 1);
 
 const {
