+++ src/pretty/README.md
@@ -1,0 +1,16 @@
+# Pretty
+
+## Pretty print to the console
+
+## Supports
+
+- Browser
+- Node
+
+## License
+
+ISC License [ISC](https://opensource.org/licenses/ISC)
+
+## Author
+
+&copy; 2018-2019 Verdexdesign
